requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-pptx>=0.6.21
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langdetect>=1.0.9
tenacity>=8.2.2
